// src/app/interfaces/catalogo-shared.ts
export interface CatalogoShared {

    tipo: string;
    iconoTituloHeader: string;
    tituloHeader: string;
    iconoBotonHeader: string;
    tituloBotonHeader: string;
    tituloTabla: Array<any>; // Viene de cada catalogo individual
    tablas: Array<any>; // Viene de cada catalogo individual
}